+++ CHANGELOG.md
@@ -29,6 +29,8 @@ it in future.
 
 ### Removed
 
+* Remove old (CDR) target configuration; Jun25 is now the only supported target
+
 * Remove unused ShipStyle, which seems copy-pasted from LHCb
 
 ## 25.12 - 2025-12-22